+++ packages/rpc/src/functions/machineTranslate.ts
@@ -1,12 +1,10 @@
-import type { Result } from "@inlang/core/utilities"
-import { telemetryNode } from "@inlang/telemetry"
 import { privateEnv } from "@inlang/env-variables"
+import type { Result } from "@inlang/core/utilities"
 
 export async function machineTranslate(args: {
 	text: string
 	targetLanguage: string
 	referenceLanguage: string
-	telemetryId?: string
 }): Promise<Result<string, Error>> {
 	try {
 		if (!privateEnv.GOOGLE_TRANSLATE_API_KEY) {
@@ -26,14 +24,9 @@ export async function machineTranslate(args: {
 		if (!response.ok) {
 			return [undefined, new Error(response.statusText)]
 		}
-		telemetryNode.capture({
-			event: "machine translation created",
-			distinctId: args.telemetryId ?? "unknown",
-		})
 		const json = await response.json()
 		return [json.data.translations[0].translatedText]
 	} catch (error) {
-		console.error("ml failed", error)
 		return [undefined, error as Error]
 	}
 }