+++ packages/sdk-js/src/adapter-sveltekit/vite-plugin/transforms/+layout.server.js.ts
@@ -2,12 +2,13 @@ import type { TransformConfig } from "../config.js"
 import { parseModule, generateCode } from "magicast"
 import { deepMergeObject } from "magicast/helpers"
 import { types } from "recast"
-import {
-	getArrowOrFunction,
-	getWrappedExport,
-	replaceOrAddExportNamedFunction,
-} from "../../../helpers/ast.js"
 import { dedent } from "ts-dedent"
+import {
+	extractWrappableExpression,
+	getWrappedExport,
+	getSdkImportedModules,
+	replaceOrAddExportNamedFunction,
+} from "../../../helpers/inlangAst.js"
 
 const requiredImports = (root: boolean) =>
 	root
@@ -32,14 +33,18 @@ export const transformLayoutServerJs = (config: TransformConfig, code: string, r
 	}
 
 	const n = types.namedTypes
-	const b = types.builders
 	const ast = parseModule(code)
 
+	// Remove imports, but save their names
+	const importNames = getSdkImportedModules(ast.$ast)
 	// Merge imports with required imports
 	const importsAst = parseModule(requiredImports(root))
 	deepMergeObject(ast, importsAst)
-	const emptyArrowFunctionDeclaration = b.arrowFunctionExpression([], b.blockStatement([]))
-	const arrowOrFunctionNode = getArrowOrFunction(ast.$ast, "load", emptyArrowFunctionDeclaration)
+	const arrowOrFunctionNode = extractWrappableExpression({
+		ast: ast.$ast,
+		name: "load",
+		availableImports: importNames,
+	})
 	const exportAst = getWrappedExport(
 		undefined,
 		[arrowOrFunctionNode],