+++ packages/sdk-js/src/adapter-sveltekit/vite-plugin/transforms/+page.svelte.ts
@@ -1,5 +1,5 @@
 import type { TransformConfig } from "../config.js"
-import { transformSvelte } from "./*.svelte.js"
+import { transformSvelte } from "./_.svelte.js"
 
 export const transformPageSvelte = (config: TransformConfig, code: string, root: boolean) => {
 	return transformSvelte(config, code)