+++ packages/sdk-js/src/adapter-sveltekit/vite-plugin/transforms/+server.js.ts
@@ -1,0 +1,6 @@
+import type { TransformConfig } from "../config.js"
+
+// TODO: implement
+export const transformServerRequestJs = (config: TransformConfig, code: string, root: boolean) => {
+	return code
+}